Improvements in films & genres (2) after testing in Postman

+++ src/main/java/ru/yandex/practicum/filmorate/model/Genre.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class Genre {
-    Integer id;
+    Short id;
     String name;
 }
