+++ src/main/java/ru/yandex/practicum/filmorate/storage/genre/GenreRowMapper.java
@@ -12,7 +12,7 @@ public class GenreRowMapper implements RowMapper<Genre> {
     @Override
     public Genre mapRow(ResultSet rs, int rowNum) throws SQLException {
         Genre genre = new Genre();
-        genre.setId(rs.getInt("id"));
+        genre.setId(rs.getShort("id"));
         genre.setName(rs.getString("name"));
         return genre;
     }
